Added Betweenable
I need it to do various checks of containment.
Because (,) is in Ord, but not the way I want it, I need to do this.

Rather than just making some MyOrd or something, though, I've instead
made it more specific.
These hold a series of methods for checking if some value is between
another two.