First Attempt at Implementing Insert
I've commented out all my hacking because I want it to stay, but I also
want to be able to build the file and test it.

I've implemented two functions which I've convinced myself I'll want in
a few places, then done an implementation of insert_node.

This impl does not make a PK-Tree, but it type-checks.
This is the first step in the process it will eventually go through, so
that's good.
It's all adding on to that from here...