Took Out a Lambda
I'm not sure about this one...
It's shorter, means the same thing, but I feel the other one was more
obvious.
Well, I guess not.
I can look at the definition and be like "Ok, it's the last two, but
flipped".

I don't know.
I guess we'll go with it.