Implemented Intersection
I had to rework Betweenable.

All of the functions it provided still exist, and can be written using
the new methods Betweenable provides, but I had to move Betweenable back
a level, because it was abstracting too much away.

With the new level, which returns which dimensions overlap, I can write
the intersect function abstractly for any Betweenable.

The name Betweenable doesn't really make sense anymore, but... meh.
Maybe I'll change that...