Fully Implemented PkTree
So, I've got some implementation for everything.
There's been very little testing so far...

find_nearest is kind of a mess.
Like, the code that's there isn't terrible, there are just a lot of
different cases that need to be handled.

I think there's more logic in find_nearest than in the entire building
and maintaining of the tree.

So, perhaps I'll revisit it later and see a really clear way to handle
it.

The tree implementation itself was ugly until this most recent rewrite,
for example, so iteration does happen.