Added Makefile
Ok, this is a general cleanup and make usefulling of the code.
So, I've added a make test rule that builds the tests and runs them.

This invloved renaming test.hs to tests.hs, so that make test wouldn't
shadow make tests, which is actually the name of the executable.

I also added targets to .gitignore.